Decide Drivetrain
Finished off Section on Deciding the Drivetrain, which encompasses decision making on the drive train type, the amount of wheels and type used and also the gearing, wheel usage and direct drive.
Changed tables in DecideDT as there were mistakes "Omni drive"
Inserted some images into BrainstormDT to give an example of differential drive in order to be consistent with other drive discussions.
Added 4 images.
Changed some links in BrainstormDT to link to the article containing the image rather than the image source.